reduce vaddr u64->u32

+++ src/core/arm/dynarmic/arm_dynarmic_32.h
@@ -54,7 +54,7 @@ public:
     void ReturnException(u32 pc, Dynarmic::HaltReason hr);
     //
     Dynarmic::CodePage cached_code_page;
-    u64 last_code_addr = u64(-1);
+    u32 last_code_addr = u32(-1);
     ArmDynarmic32& m_parent;
     Core::Memory::Memory& m_memory;
     Kernel::KProcess* m_process{};
